statistics: add more test for tidb_merge_partition_stats_concurrency
